count FPS across scene changes
