+++ UltraStar Play/Assets/Common/Scene/CommonSceneObjects.prefab
@@ -680,6 +680,7 @@ Transform:
   - {fileID: 9143355034012100345}
   - {fileID: 2908706230957184142}
   - {fileID: 8570025457633197298}
+  - {fileID: 8635082549009352342}
   m_Father: {fileID: 8069444317427349030}
   m_RootOrder: -1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -6693,6 +6694,51 @@ MonoBehaviour:
     type: 3}
   contextMenuSeparatorUi: {fileID: 9197481963319205126, guid: 95883c4a81ebaf142a628840e34b730f,
     type: 3}
+--- !u!1 &8870676065052498659
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 8635082549009352342}
+  - component: {fileID: 2522791573740333623}
+  m_Layer: 0
+  m_Name: ShowFpsControl
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &8635082549009352342
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8870676065052498659}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 5483831765678408185}
+  m_RootOrder: -1
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &2522791573740333623
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8870676065052498659}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 84c038ec799bea549b582079d30ad6b0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  fps: 0
 --- !u!1 &9038781141353725927
 GameObject:
   m_ObjectHideFlags: 0
