--- UltraStar Play/Packages/playshared/Runtime/UI/Fps/ShowFpsText.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 50d02acac4813a94490c57776dc5e4ee
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
